Fix review issues: SSE injection, UUID error handling, zero-dep core
- Sanitize SSE event name/id fields to prevent newline injection
- Panic on crypto/rand failure in UUID generation (fail-fast)
- Move Sonic JSON to separate Go module (json/go.mod) — core is zero-dep
- Use encoding/json as default, Sonic opt-in via WithJSONEncoder
- Security headers respect user-set values (no overwrite)
- CORS sets Vary: Origin on rejected origins (CDN caching fix)
- parseSize returns int64 to prevent overflow on 32-bit systems
- Example checks Listen() error

+++ validation_pbt_test.go
@@ -114,7 +114,7 @@ func TestPropertyValidateMaxSizeAgreesWithParseSize(t *testing.T) {
 
 		fu := &FileUpload{Size: in.FileSize}
 		got := validateMaxSize(fu, param)
-		expected := in.FileSize <= int64(maxBytes)
+		expected := in.FileSize <= maxBytes
 		return got == expected
 	}
